Speakers are informed of application approval and other fixes
'reviews remaining' counter on reviewer dashboard now updated properly
